requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
astroplan==0.6
astropy==4.0
click==7.0                # via flask
cycler==0.10.0            # via matplotlib
decorator==4.4.1          # via networkx
flask==1.1.1
imageio==2.6.1            # via scikit-image
itsdangerous==1.1.0       # via flask
jinja2==2.11.1            # via flask
kiwisolver==1.1.0         # via matplotlib
loguru==0.4.1
markupsafe==1.1.1         # via jinja2
matplotlib==3.1.3
networkx==2.4             # via scikit-image
numpy==1.18.1
photutils==0.7.2
pillow==7.0.0             # via imageio, scikit-image
pyparsing==2.4.6          # via matplotlib
pyserial==3.4
python-dateutil==2.8.1
python-json-logger==0.1.11
pytz==2019.3              # via astroplan
pywavelets==1.1.1         # via scikit-image
pyyaml==5.3
pyzmq==18.1.1
ruamel.yaml.clib==0.2.0   # via ruamel.yaml
ruamel.yaml==0.16.10
scalpl==0.3.0
scikit-image==0.16.2
scipy==1.4.1
six==1.14.0               # via astroplan, cycler, python-dateutil
versioneer==0.18
werkzeug==1.0.0           # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools==45.2.0        # via kiwisolver
